dimw has been renamed to dimworld
The overall dimworld of a coupling problem is now the maximum
of the dimworlds of the individual grids.  This is necessary
for true 1d-2d couplings.

I can even think of cases where this is not general enough
(contact between two 2d shells in 3d), and hence eventually
we may want to have dimworld as a user parameter.  But for
the time being the maximum is okay.


[[Imported from SVN: r224]]