Replace brute-force quadratic algorithm by the linear one by Gander and Japhet.
The new algorithm combines two advancing fronts and has expected linear time.
This is optimal because in the worst case you may have a quadratic number of
intersections.  The original algorithm was described in 

'An Algorithm for Non-Matching Grid Projections with Linear Complexity,
 M.J. Gander and C. Japhet, Domain Decomposition Methods in Science and Engineering XVIII, 
pp. 185--192, Springer-Verlag, 2009.'

StandardMerge now implements a simplified variant of this.

The code is still in a prototype stage.  In particular it still has quadradic
run-time.  I will clean it up in the next few days.

[[Imported from SVN: r452]]