Get rid of the variable 'parents'
It was counting the size of an array, so we may as well
use the array size directly.

Note: the GridGlueVtkWriter doesn't compile at all with the
current interface, and the transition is not quite
straightforward.  I'll try to make it step by step, but
can't actually test in between.


[[Imported from SVN: r597]]