For the new parallelization strategy we will call the merger several times. ATM we assume that the merger always holds all necessary information.
This patch changes most places where the merger is needed after the
setup phase.

Nice plus: Once we added a clear() method to the merger we can safe
quite some memory, as no duplicate data has to be stored. All final
information are stored in the IntersectionData objects or in the
patch.

[[Imported from SVN: r608]]