Allow the computeIntersection to return information whether neighboring elements would also intersect.
This is used by Gander and Japhet to reduce the number of calls to the computeIntersection
routine.  If you hand-code your computeIntersection the information is cheaply available.
It is not quite as easy with cgal, but should still be worth it.

This patch is only the method signature change.  The information is not computed yet,
and it is not used yet.

[[Imported from SVN: r640]]