Actually fill the neighborIntersects fields.
So far only have way:  if a field contains a 'true' that means
that the neighbor across that face will also intersect.  If
it contains 'false' that means 'don't know'.

So far this is a mixed blessing.  In my (small) test case the
time gained by calculating fewer intersections is pretty much
offset by the time needed for the extra information.

[[Imported from SVN: r663]]