Remove CGALMerge completely.
It works very nicely, but due to the use of exact arithmetic
it is unbearably slow.  I tried fairly hard about a year ago
to make it faster, in particular to get it to run with a
faster number type, but to no avail.  Now we have the new
OverlappingMerge implementation which does the same thing,
and is much faster.  Hence CGALMerge is dropped.

Merry Christmas :-)

[[Imported from SVN: r741]]