[intersection] add new method flip() to get an itersection from the neighboring domain
With Andreas I have discussed different options to fix the problem of
non-symmetric intersections. For him this problem occured with surface
meshes, but we have a similar problem with grid-glue -- in our case
the geometric information is the same, but we have different types for
the two sides. One proposed solution was a method flip(), which
returns the same intersection from the neighboring cell. Now I add
this method to allow us to get some first experience.