[indexset] add new IntersectionIndexSet
we used a quickhack to provide access to somewthing like an IntersectionIndexSet. This patch
replaces the hack with a proper IndexSet implementation. The GridGlue now has a method
indexSet() which returns an object and the couplingtest now tests the IndexSet, to make sure 
the index is zero starting and consecutive.