Extract boundaries of 2d grids without using orientedSubface
The orientedSubface mechanism is notoriously unreliable.  It works for
some grids, and for others it doesn't.  For 2d grids at least, we now
orient the extracted boundary faces by comparing their normals with
the intersection normal returned by centerUnitOuterNormal.  That should
work in all cases.

The corresponding patch for 3d grids will follow in a separate patch.