inexactInverseProjection: use a better initial guess for distance (x[2])
An initial guess of 1/3 for the distance between `target' and its
preimage can be arbitrarily wrong as the correct distance depends on
the scale of the problem.

This patch takes instead the distance between the target and the
center of the triangle which is already used as the initial value for
the local coordinates (x[0], x[1]).