Inverse projection now returns distance along normal at image corner
The inverse projection now returns the distance along the normal n(yᵢ)
at the image corner yᵢ instead of the normal n(Φ⁻¹(yᵢ)) on the preimage
triangle.

With the change projectionFeasible() actually checks the distance along
both normals as inteded. Before it did check the same distance twice,
resulting in invalid projections to be considered valid.