inside: if an entry is NaN, return false
If any of the xᵢ is NaN, the point is invalid and certainly not inside
the standard simplex. As in this case "sum" will also be NaN, any
comparison involving "sum" will be false. We make use of that to catch
this case by a small change in the comparison logic (which seems
equivalent at first glance).