Make grid dimensions easier to change
This patch makes the grid dimensions easier to change for future
testing of 2d/3d couplings which has additional interesting problems
that do not exist for 1d/2d couplings.

The `MixedDimTrafo` still assumes it is used to lift a 1d grid into 2d
space and needs to be updated later.