Add `.gitlab-ci.yml` for 2.4 release branch
The 2.4 release branch should work with the 2.3 and 2.4 releases of
the DUNE core modules, with both autotools and cmake.  We also test
one configuration with clang instead of the default compiler (gcc).

Note that `make install` still needs to be tested manually and
compared between the CMake and autotools build systems.