Dismiss projection if forward projection is too far in wrong direction
If the forward projection of a vertex can only be achieved by projecting
in the wrong projection direction, the forward projected triangle
should not be used for an inverse projection or edge intersection, as
the results are usually wrong.
One exception is the case of initially penetrating grids, where one might
still want to keep the resulting intersection. For this case the projection
is dismissed if the forward projection is further (in the wrong direction)
than two times the allowed overlap.