Add possibility to switch between closest point and normal projection
In contact problems both projections are commonly used.
For the outer normal projection the projection is done along the outer
normals of the grid1. The closest point projection is done along the
outer normal of grid2, due to the projection being a best approximation.
Hence both projections can be implemented via a simple switch of ordering.